requirements.txt
transformers
streamlit
streamlit-chat
sentence-transformers
pinecone-client
PyMuPDF
langchain
langchain-community
openai
torch 
sentence_transformers 
faiss-cpu
pypdf
langchain-pinecone
groq